+++ pkg/db/model.go
@@ -3,16 +3,17 @@ package db
 // RedboxAccount is a type corresponding to a RedboxAccount table record
 type RedboxAccount struct {
 	ID             string        `json:"Id"`             // AWS Account ID
-	GroupID        string        `json:"GroupId"`        // Azure AD Group ID associated with the Account
 	AccountStatus  AccountStatus `json:"AccountStatus"`  // Status of the AWS Account
 	LastModifiedOn int64         `json:"LastModifiedOn"` // Last Modified Epoch Timestamp
+	CreatedOn      int64         `json:"CreatedOn"`
+	AdminRoleArn   string        `json:"AdminRoleArn"` // Assumed by the Redbox master account, to manage this user account
 }
 
 // RedboxAccountAssignment is a type corresponding to a RedboxAccountAssignment
 // table record
 type RedboxAccountAssignment struct {
 	AccountID        string           `json:"AccountId"`        // AWS Account ID
-	UserID           string           `json:"UserId"`           // Azure User ID
+	UserID           string           `json:"UserId"`
 	AssignmentStatus AssignmentStatus `json:"AssignmentStatus"` // Status of the Assignment
 	CreatedOn        int64            `json:"CreatedOn"`        // Created Epoch Timestamp
 	LastModifiedOn   int64            `json:"LastModifiedOn"`   // Last Modified Epoch Timestamp