+++ pkg/provision/mocks/Provisioner.go
@@ -11,17 +11,26 @@ type Provisioner struct {
 }
 
 // ActivateAccountAssignment provides a mock function with given fields: _a0, _a1, _a2
-func (_m *Provisioner) ActivateAccountAssignment(_a0 bool, _a1 string, _a2 string) error {
+func (_m *Provisioner) ActivateAccountAssignment(_a0 bool, _a1 string, _a2 string) (*db.RedboxAccountAssignment, error) {
 	ret := _m.Called(_a0, _a1, _a2)
 
-	var r0 error
-	if rf, ok := ret.Get(0).(func(bool, string, string) error); ok {
+	var r0 *db.RedboxAccountAssignment
+	if rf, ok := ret.Get(0).(func(bool, string, string) *db.RedboxAccountAssignment); ok {
 		r0 = rf(_a0, _a1, _a2)
 	} else {
-		r0 = ret.Error(0)
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(*db.RedboxAccountAssignment)
+		}
 	}
 
-	return r0
+	var r1 error
+	if rf, ok := ret.Get(1).(func(bool, string, string) error); ok {
+		r1 = rf(_a0, _a1, _a2)
+	} else {
+		r1 = ret.Error(1)
+	}
+
+	return r0, r1
 }
 
 // FindUserActiveAssignment provides a mock function with given fields: _a0